overloaded calculator methods

+++ src/main/java/org/fasttrackit/Calculator.java
@@ -1,13 +1,14 @@
 package org.fasttrackit;
 
 public class Calculator {
-    public int sum(int a, int b) {
+    public double sum(double a, double b) {
         return a + b;
     }
-    public int multiply(int a, int b) {
+    public double multiply(double a, double  b) {
         return a * b;
     }
     public double divide(double a, double b) {
+
         return a / b;
     }
 }
